+++ pkg/deployments/CHANGELOG.md
@@ -41,6 +41,7 @@
 
 - Deployed all L2 contracts to Avalanche.
 - Deployed all contracts on Goerli to Sepolia.
+- Deployed all L2 contracts to Polygon zkEVM.
 
 ### New Deployments